removing unecessary helper classes
