+++ src/api/java/leaf/cosmere/api/Manifestations.java
@@ -15,7 +15,7 @@ public class Manifestations
 {
 	public final static int ALLOMANCY_ID = 1;
 	public final static int FERUCHEMY_ID = 2;
-	public final static int HEMALURGY_ID = 3;//yes I know. Didn't add this till way later and now I'm too afraid to change it... todo: change it
+	public final static int HEMALURGY_ID = 3;//yes I know. Didn't add this till way later and now I'm too afraid to change it... I did what needed to be done
 	public final static int SURGEBINDING_ID = 4;
 	public final static int AONDOR_ID = 5;
 	public final static int AWAKENING_ID = 6;
